[Add] 음식 태그 보여주는 ScrollView 하단 포커싱 추가 및 뷰 분리 #45
