+++ FinalHrmiProjects/View/Record/FoodTagAddTextField.swift
@@ -1,0 +1,49 @@
+//
+//  FoodTagAddTextField.swift
+//  FinalHrmiProjects
+//
+//  Created by 정인선 on 1/30/24.
+//
+
+import SwiftUI
+
+struct FoodTagAddTextField: View {
+    // 음식 태그 배열
+    @Binding var foodTags: [FoodTag]
+    // TextField로 부터 입력받은 음식 태그 이름
+    @State private var foodTagName: String = ""
+    
+    var body: some View {
+        HStack {
+            Text("#")
+                .font(.regular16)
+                .opacity(0.7)
+            
+            TextField("음식 이름", text: $foodTagName)
+            Button {
+                // 중복 추가 불가
+                if !foodTagName.isEmpty && !foodTags.contains(where: { $0.name == foodTagName }) {
+                    foodTags.append(FoodTag(name: foodTagName))
+                }
+                // TextField 비워주기
+                foodTagName = ""
+            } label: {
+                Text("추가하기")
+                    .font(.regular14)
+                    .opacity(0.7)
+                
+            }
+            // tint color 안 들어가게 버튼 스타일 변경
+            .buttonStyle(.plain)
+        }
+        .padding(.horizontal, 20)
+        .padding(.vertical, 10)
+        .background(.mainAccent03.opacity(0.2))
+        .clipShape(.rect(cornerRadius: 10))
+        .padding(.horizontal, 20)
+    }
+}
+
+//#Preview {
+//    FoodTagAddTextField()
+//}
